added PlayerHealthController
