+++ Mad-Drake/Assets/Scripts/HealthBar.cs
@@ -8,6 +8,7 @@ public class HealthBar : MonoBehaviour
     public Slider slider;
     public Gradient gradient;
     public Image fill;
+
     public void SetMaxHealth(int health)
     {
         slider.maxValue = health;
