+++ Mad-Drake/Assets/Scripts/PlayerHealthController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5cd5003f0f2bc5b4dbe28d9013bba8cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
